feat: 이미지 업로드 형식에 맞게 수정, 리사이즈 추상화
